+++ src/etoolbox/utils/misc.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import requests
 
-from etoolbox.utils.lazy_import import lazy_import
+from etoolbox._optional import logging_redirect_tqdm, tqdm
 
 
 def download(url: str, fname: Path | BytesIO) -> None:
@@ -17,16 +17,13 @@ def download(url: str, fname: Path | BytesIO) -> None:
     Returns: None
 
     """
-    tqdm = lazy_import("tqdm.auto", wait_for_signal=False)
-    tqdm_logging = lazy_import("tqdm.contrib.logging", wait_for_signal=False)
-
     resp = requests.get(url, stream=True)  # noqa: S113
     total = int(resp.headers.get("content-length", 0))
     # Can also replace 'file' with an io.BytesIO object
     with (
-        tqdm_logging.logging_redirect_tqdm(),
+        logging_redirect_tqdm(),
         open(fname, "wb") as file,
-        tqdm.tqdm(
+        tqdm(
             desc="Downloading " + fname.name,
             total=total,
             unit="iB",
@@ -52,17 +49,14 @@ def ungzip(zip_path: Path, out_path=None):
     """
     import gzip
 
-    tqdm = lazy_import("tqdm.auto", wait_for_signal=False)
-    tqdm_logging = lazy_import("tqdm.contrib.logging", wait_for_signal=False)
-
     if out_path is None:
         out_path = zip_path.parent / zip_path.stem
 
     with (
-        tqdm_logging.logging_redirect_tqdm(),
+        logging_redirect_tqdm(),
         gzip.open(zip_path, "rb") as zf,
         open(out_path, "wb") as out,
-        tqdm.tqdm(
+        tqdm(
             desc="Ungzipping " + zip_path.name,
             unit="iB",
             unit_scale=True,